Update unit tests

+++ tests/unit/metafunctions/varybivore/is_different_from.test.hpp
@@ -4,8 +4,8 @@
 #ifndef CONCEPTRODON_TESTS_UNIT_VARYBIVORE_IS_DIFFERENT_FROM_H
 #define CONCEPTRODON_TESTS_UNIT_VARYBIVORE_IS_DIFFERENT_FROM_H
 
-#include <concepts>
 #include "conceptrodon/varybivore/is_different_from.hpp"
+
 #include "macaron/judgmental/valid.hpp"
 #include "macaron/judgmental/invalid.hpp"
 
@@ -23,7 +23,7 @@ namespace TestIsDifferent {
 /******************************************************************************************************/
 INVALID(IsDifferentFrom<1>::Page_v<1>);
 VALID(IsDifferentFrom<1>::Page<2>::value);
-VALID(IsDifferentFrom<1>::Page_v<static_cast<size_t>(1)>);
+VALID(IsDifferentFrom<1>::Page_v<1.0>);
 /******************************************************************************************************/
