+++ tests/unit/metafunctions/varybivore/maximum.test.hpp
@@ -4,8 +4,8 @@
 #ifndef CONCEPTRODON_TESTS_UNIT_VARYBIVORE_MAXIMUM_H
 #define CONCEPTRODON_TESTS_UNIT_VARYBIVORE_MAXIMUM_H
 
-#include <concepts>
 #include "conceptrodon/varybivore/maximum.hpp"
+
 #include "macaron/judgmental/equal_value.hpp"
 
 #include "macaron/judgmental/amenity/define_equal_value.hpp"
@@ -21,9 +21,11 @@ namespace TestMaximum {
 /******************************************************************************************************/
 #define SUPPOSED_VALUE  \
     12
+
 EQUAL_VALUE(Maximum_v<-10, 11, 12>);
 EQUAL_VALUE(Maximum<-10, 11, 12>::value);
-EQUAL_VALUE(Maximum_v<-10.0, -9.0, -8.5, static_cast<double>(12)>);
+EQUAL_VALUE(Maximum_v<-10.0, -9.0, -8.5, 12.0>);
+
 #undef SUPPOSED_VALUE
 /******************************************************************************************************/
